Rename file: const -> spec
