+++ jam_types/spec.py
@@ -42,15 +42,15 @@ def set_spec(spec_name):
     if spec_name not in SPECS:
         raise ValueError(f"Unknown spec: {spec_name}. Available specs: {list(SPECS.keys())}")
     _current_spec = spec_name
-    _update_constants()
+    _update_globals()
     _update_type_classes()
 
 def get_current_spec():
     """Get the name of the current spec."""
     return _current_spec
 
-def _update_constants():
-    """Update all derived constants based on current spec."""
+def _update_globals():
+    """Update all derived globals based on current spec."""
     global validators_count, epoch_length, max_tickets_per_block
     global validators_per_core, core_count, validators_super_majority, avail_bitfield_bytes
     global auth_pool_max_size, auth_queue_size, hash_size
@@ -58,18 +58,18 @@ def _update_constants():
     logging.debug("Using JAM spec: %s", _current_spec)
     spec = SPECS[_current_spec]
     
-    # Base constants from spec
+    # Base globals from spec
     validators_count = spec['validators_count']
     epoch_length = spec['epoch_length']
     max_tickets_per_block = spec['max_tickets_per_block']
     
-    # Derived constants
+    # Derived values
     validators_per_core = 3
     core_count = validators_count // validators_per_core
     validators_super_majority = validators_count // 3 * 2 + 1
     avail_bitfield_bytes = (core_count + 7) // 8
     
-    # Fixed constants
+    # Fixed values
     auth_pool_max_size = 8
     auth_queue_size = 80
     hash_size = 32
@@ -77,9 +77,9 @@ def _update_constants():
 def _update_type_classes():
     """Update all registered spec-dependent classes."""
     for cls, spec_attrs in _spec_dependent_classes:
-        for attr_name, const_name in spec_attrs.items():
-            logging.debug("Updating {} {} = {}".format(cls, attr_name, globals()[const_name]))
-            setattr(cls, attr_name, globals()[const_name])
+        for attr_name, glob_name in spec_attrs.items():
+            logging.debug("Updating {} {} = {}".format(cls, attr_name, globals()[glob_name]))
+            setattr(cls, attr_name, globals()[glob_name])
 
-# Initialize constants with current spec
-_update_constants()
+# Initialize globals with current spec
+_update_globals()
